dev-requirements.txt
attrs==23.2.0
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
defusedxml==0.7.1
distlib==0.3.8
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.15.1
h11==0.14.0
hyperlink==21.0.0
identify==2.5.36
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==5.2.2
nodeenv==1.9.1
openpyxl==3.1.4
outcome==1.3.0.post0
packaging==24.1
parsel==1.9.1
platformdirs==4.2.2
-e git+ssh://git@github.com/jetale/postings_parser.git@5123b5bd714070b2552e78bbd47f1a6019a2b953#egg=postings_parser
pre-commit==3.7.1
Protego==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
selenium==4.21.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tldextract==5.1.2
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
typing_extensions==4.12.2
urllib3==2.2.1
virtualenv==20.26.3
w3lib==2.2.1
wsproto==1.2.0
zope.interface==6.4.post2
